feat(ban): Marshal/unmarshal ban data in ban reason for richer data.
This commit makes BanHanderData available outside of the ban
interaction package, so that the ban listener may also make use of it.
A method and function has been added to marshal/stringify the structured
data into Discord's ban reason, and then later unmarshal/parse it back
into structured data.

The first part of the message is the ban reason. In this way, any
unparseable data just becomes part of the reason. Then, there is a new
line followed by three instances of INFORMATION SEPARATOR ONE (U+001F)
on its own line, and finally, newline-separated key/value pairs, similar
to that in Git commits. Keys and values are separated by a colon.