fix: Minor changes to modmail UX.
This commit removes the initial "||" before mentions in the message
starting a Modmail thread, adds a minimum description length for
reports, and removes it for message reports. It also manually sets the
"slow-mode-time" parameter to "0s" if not set, which has no effect for
the end-user, but prevents a parse error from occurring.