feat: Added ban footer for ban DMs
Added a new "ban footer", which is a moderator-set text that is included
when a ban DM is sent to the user. It can also be set to send it, even
if there is no ban message or ban duration specified. The intended use
is to be able to inform banned users of appeal policies and similar.

In the cases one wants to avoid informing people, when it sends out
automatically, there is a new option in the ban command. Note that
adding a ban message or setting a duration will still cause a DM to be
sent, with the footer.

Adds a new "/admin ban-footer" sub-command. No parameters will result in
current configuration being shown, with a button to open a modal to edit
the message.
Modifies the "/ban" command, adding a new, optional parameter
"dont-auto-dm".